feat: added new controller to handle ongoing purchase services
