+++ api-gateway/src/main/resources/application.properties
@@ -21,6 +21,7 @@ spring.cloud.gateway.routes[0].predicates[0]=Path=\
   /api/v1/movies/*, \
   /api/v1/payments, \
   /api/v1/payments/**,\
+  /api/v1/reserves/**,\
   /api/v1/schedules/*, \
   /api/v1/demo/*
