+++ movie-service/src/main/java/com/barbenheimer/movieservice/controller/PurchaseController.java
@@ -22,36 +22,11 @@ public class PurchaseController {
 
     private final PurchaseService purchaseService;
 
-    private OngoingPurchaseService ongoingPurchaseService;
-
     @Autowired
-    public PurchaseController(PurchaseService purchaseService, OngoingPurchaseService ongoingPurchaseService){
+    public PurchaseController(PurchaseService purchaseService){
         this.purchaseService = purchaseService;
-        this.ongoingPurchaseService = ongoingPurchaseService;
     }
 
-    /**
-     * This method gets an ongoing purchase's details with its token.
-     * @param paymentIntentId
-     * @return ResponseEntity<OngoingPurchaseDetailDTO>
-     */
-    @GetMapping("/{paymentIntentId}")
-    public ResponseEntity<OngoingPurchaseDetailDTO> getOngoingPurchaseDetail(@PathVariable String paymentIntentId){
-        return ResponseEntity.ok(ongoingPurchaseService.getDetail(paymentIntentId));
-    }
-
-
-    /**
-     * This method checks if an ongoing purchase is still valid.
-     * @param paymentIntentId
-     * @return ResponseEntity<?>
-     * @throws StripeException
-     */
-    @GetMapping("/checkToken/{paymentIntentId}")
-    public ResponseEntity<?> checkIfValidToken(@PathVariable String paymentIntentId) throws StripeException {
-        PaymentIntentValidationDTO paymentIntentValidationDTO = ongoingPurchaseService.checkIfValidToken(paymentIntentId);
-        return ResponseEntity.ok(paymentIntentValidationDTO);
-    }
 
     /**
      * This method gets the purchase linked to a payment intent id for the frontend to display its details on a payment summary page.
